pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "chuang_tzu_bot"
version = "0.1.0"
description = ""
requires-python = ">=3.12"

dependencies = [
    "aiogram",
    "python-dotenv",
    "openai",
    "newspaper_boy @ git+https://github.com/rheophile10/msm.git",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/chuang_tzu_bot"]
